Move snapshot client to pkg
