+++ cmd/vcluster/cmd/snapshot/list.go
@@ -15,7 +15,7 @@ func NewListCmd() *cobra.Command {
 		Short: "list vCluster snapshots",
 		Args:  cobra.NoArgs,
 		RunE: func(cmd *cobra.Command, _ []string) error {
-			client := &Client{}
+			client := &snapshot.Client{}
 			envOptions, err := snapshot.ParseOptionsFromEnv()
 			if err != nil {
 				return fmt.Errorf("failed to parse options from environment: %w", err)
