+++ cmd/vcluster/cmd/snapshot/snapshot.go
@@ -1,25 +1,11 @@
 package snapshot
 
 import (
-	"archive/tar"
-	"compress/gzip"
-	"context"
-	"encoding/json"
 	"fmt"
-	"os"
 
-	"github.com/loft-sh/vcluster/pkg/config"
-	"github.com/loft-sh/vcluster/pkg/constants"
-	"github.com/loft-sh/vcluster/pkg/etcd"
 	"github.com/loft-sh/vcluster/pkg/snapshot"
-	"github.com/loft-sh/vcluster/pkg/snapshot/types"
 	"github.com/spf13/cobra"
-	"k8s.io/klog/v2"
 )
-
-type Client struct {
-	Options snapshot.Options
-}
 
 func NewSnapshotCommand() *cobra.Command {
 	cmd := &cobra.Command{
@@ -27,7 +13,7 @@ func NewSnapshotCommand() *cobra.Command {
 		Short: "snapshot a vCluster",
 		Args:  cobra.NoArgs,
 		RunE: func(cmd *cobra.Command, _ []string) error {
-			client := &Client{}
+			client := &snapshot.Client{}
 			envOptions, err := snapshot.ParseOptionsFromEnv()
 			if err != nil {
 				return fmt.Errorf("failed to parse options from environment: %w", err)
@@ -42,191 +28,4 @@ func NewSnapshotCommand() *cobra.Command {
 	cmd.AddCommand(NewDeleteCmd())
 
 	return cmd
-}
-
-func (c *Client) Run(ctx context.Context) error {
-	// parse vCluster config
-	vConfig, err := config.ParseConfig(constants.DefaultVClusterConfigLocation, os.Getenv("VCLUSTER_NAME"), nil)
-	if err != nil {
-		return err
-	}
-
-	// make sure to validate options
-	err = snapshot.ValidateConfigAndOptions(vConfig, &c.Options, false, false)
-	if err != nil {
-		return err
-	}
-
-	// create new etcd client
-	etcdClient, err := snapshot.NewEtcdClient(ctx, vConfig, false)
-	if err != nil {
-		return fmt.Errorf("failed to create etcd client: %w", err)
-	}
-
-	// create store
-	objectStore, err := snapshot.CreateStore(ctx, &c.Options)
-	if err != nil {
-		return fmt.Errorf("failed to create store: %w", err)
-	}
-
-	// write the snapshot
-	klog.Infof("Start writing etcd snapshot %s...", objectStore.Target())
-	err = c.writeSnapshot(ctx, etcdClient, objectStore)
-	if err != nil {
-		return err
-	}
-
-	klog.Infof("Successfully wrote snapshot to %s", objectStore.Target())
-	return nil
-}
-
-func (c *Client) List(ctx context.Context) ([]types.Snapshot, error) {
-	// parse vCluster config
-	vConfig, err := config.ParseConfig(constants.DefaultVClusterConfigLocation, os.Getenv("VCLUSTER_NAME"), nil)
-	if err != nil {
-		return nil, err
-	}
-
-	// make sure to validate options
-	err = snapshot.ValidateConfigAndOptions(vConfig, &c.Options, false, true)
-	if err != nil {
-		return nil, err
-	}
-
-	// create store
-	objectStore, err := snapshot.CreateStore(ctx, &c.Options)
-	if err != nil {
-		return nil, fmt.Errorf("failed to create store: %w", err)
-	}
-
-	// list snapshots
-	return objectStore.List(ctx)
-}
-
-func (c *Client) Delete(ctx context.Context) error {
-	// parse vCluster config
-	vConfig, err := config.ParseConfig(constants.DefaultVClusterConfigLocation, os.Getenv("VCLUSTER_NAME"), nil)
-	if err != nil {
-		return err
-	}
-
-	// make sure to validate options
-	err = snapshot.ValidateConfigAndOptions(vConfig, &c.Options, false, false)
-	if err != nil {
-		return err
-	}
-
-	// create store
-	objectStore, err := snapshot.CreateStore(ctx, &c.Options)
-	if err != nil {
-		return fmt.Errorf("failed to create store: %w", err)
-	}
-
-	// delete snapshot
-	if err := objectStore.Delete(ctx); err != nil {
-		return err
-	}
-
-	klog.Infof("Successfully deleted snapshot %s", objectStore.Target())
-	return nil
-}
-
-func (c *Client) writeSnapshot(ctx context.Context, etcdClient etcd.Client, objectStore types.Storage) error {
-	// now stream objects from etcd to object store
-	errChan := make(chan error)
-	reader, writer, err := os.Pipe()
-	if err != nil {
-		return fmt.Errorf("failed to create pipe: %w", err)
-	}
-	defer writer.Close()
-	go func() {
-		defer reader.Close()
-		errChan <- objectStore.PutObject(ctx, reader)
-	}()
-
-	// start listing the keys
-	listChan := etcdClient.ListStream(ctx, "/")
-
-	// don't compress with default level as this can get quite cpu heavy
-	gzipWriter, _ := gzip.NewWriterLevel(writer, 3)
-	defer gzipWriter.Close()
-
-	// create a new tar write
-	tarWriter := tar.NewWriter(gzipWriter)
-	defer tarWriter.Close()
-
-	// write the vCluster config as first thing
-	if c.Options.Release != nil {
-		releaseBytes, err := json.Marshal(c.Options.Release)
-		if err != nil {
-			return fmt.Errorf("failed to marshal vCluster release: %w", err)
-		}
-
-		err = writeKeyValue(tarWriter, []byte(snapshot.SnapshotReleaseKey), releaseBytes)
-		if err != nil {
-			return fmt.Errorf("failed to snapshot vCluster release: %w", err)
-		}
-	}
-
-	// now write the snapshot
-	backedUpKeys := 0
-	for {
-		select {
-		case <-ctx.Done():
-			return fmt.Errorf("context: %w", ctx.Err())
-		case err := <-errChan:
-			if err != nil {
-				return fmt.Errorf("failed to write snapshot: %w", err)
-			}
-			return nil
-		case obj := <-listChan:
-			// check if error or object to write
-			if obj != nil {
-				if obj.Error != nil {
-					return fmt.Errorf("failed to retrieve etcd items: %w", obj.Error)
-				}
-
-				// write the object into the store
-				klog.V(1).Infof("Snapshot key %s", string(obj.Value.Key))
-				err := writeKeyValue(tarWriter, obj.Value.Key, obj.Value.Data)
-				if err != nil {
-					return fmt.Errorf("failed to snapshot key %s: %w", string(obj.Value.Key), err)
-				}
-
-				// print status update
-				backedUpKeys++
-				if backedUpKeys%100 == 0 {
-					klog.Infof("Backed up %d keys", backedUpKeys)
-				}
-			} else {
-				klog.Infof("Successfully backed up %d etcd keys", backedUpKeys)
-
-				// close the writer to signal we are done, but wait until object store has finished writing
-				_ = tarWriter.Close()
-				_ = gzipWriter.Close()
-				_ = writer.Close()
-				return <-errChan
-			}
-		}
-	}
-}
-
-func writeKeyValue(tarWriter *tar.Writer, key, value []byte) error {
-	err := tarWriter.WriteHeader(&tar.Header{
-		Typeflag: tar.TypeReg,
-		Name:     string(key),
-		Size:     int64(len(value)),
-		Mode:     0666,
-	})
-	if err != nil {
-		return err
-	}
-
-	// write value to tar archive
-	_, err = tarWriter.Write(value)
-	if err != nil {
-		return err
-	}
-
-	return nil
 }
